refactor : 클래스 역할에 맞게 분리
- group 내에서 각 기능별로 따로 클래스를 만들어 분리함
