+++ src/main/java/com/alal/backend/controller/user/GroupController.java
@@ -6,8 +6,9 @@ import com.alal.backend.domain.dto.request.UploadProjectRequest;
 import com.alal.backend.domain.dto.request.UploadSceneRequest;
 import com.alal.backend.domain.dto.response.*;
 import com.alal.backend.service.group.ProjectService;
-import com.alal.backend.service.user.AvatarService;
-import com.alal.backend.service.group.GroupService;
+import com.alal.backend.service.group.SceneService;
+import com.alal.backend.service.group.AvatarService;
+import com.alal.backend.service.group.MemoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
@@ -20,10 +21,10 @@ import java.io.IOException;
 @RequestMapping("/group")
 @RequiredArgsConstructor
 public class GroupController {
-
-    private final GroupService groupService;
+    private final MemoService memoService;
     private final AvatarService avatarService;
     private final ProjectService projectService;
+    private final SceneService sceneService;
 
     @PostMapping("/memo/upload")
     public ResponseEntity<UploadMemoResponse> upload(@RequestBody UploadMemoRequest uploadMemoRequest
@@ -31,7 +32,7 @@ public class GroupController {
     ) {
         Long userId = 1L;
 //        Long userId = userPrincipal.getId();
-        return ResponseEntity.ok(groupService.uploadMemo(uploadMemoRequest, userId));
+        return ResponseEntity.ok(memoService.uploadMemo(uploadMemoRequest, userId));
     }
 
     @GetMapping("/memo")
@@ -40,7 +41,7 @@ public class GroupController {
     ) {
         Long userId = 1L;
 //        Long userId = userPrincipal.getId();
-        return ResponseEntity.ok(groupService.readMemos(userId));
+        return ResponseEntity.ok(memoService.readMemos(userId));
     }
 
     @PostMapping("/project/upload")
@@ -49,7 +50,7 @@ public class GroupController {
                                                                ) {
         Long userId = 1L;
 //        Long userId = userPrincipal.getId();
-        return ResponseEntity.ok(groupService.uploadProject(uploadProjectRequest, userId));
+        return ResponseEntity.ok(projectService.uploadProject(uploadProjectRequest, userId));
     }
 
     @GetMapping("/project/all")
@@ -58,12 +59,12 @@ public class GroupController {
             ) {
 //        Long userId = userPrincipal.getId();
         Long userId = 1L;
-        return ResponseEntity.ok(groupService.readProjects(userId, pageable));
+        return ResponseEntity.ok(memoService.readProjects(userId, pageable));
     }
 
     @GetMapping("/project/{id}")
     public ResponseEntity<ReadProjectResponse> readProject(@PathVariable("id") Long projectId) throws IOException {
-        ReadProjectResponse readProjectResponse = groupService.readProject(projectId);
+        ReadProjectResponse readProjectResponse = memoService.readProject(projectId);
         return ResponseEntity.ok(readProjectResponse);
     }
 
@@ -74,11 +75,11 @@ public class GroupController {
 
     @GetMapping("/scene/{id}")
     public ResponseEntity<ReadSceneResponseList> readAll(@PathVariable("id") Long projectId) {
-        return ResponseEntity.ok(projectService.readAllScene(projectId));
+        return ResponseEntity.ok(sceneService.readAllScene(projectId));
     }
 
     @PostMapping("/scene/upload")
     public ResponseEntity<UploadSceneResponse> upload(@RequestBody UploadSceneRequest uploadSceneRequest) {
-        return ResponseEntity.ok(projectService.uploadScene(uploadSceneRequest));
+        return ResponseEntity.ok(sceneService.uploadScene(uploadSceneRequest));
     }
 }
