+++ src/main/java/com/alal/backend/repository/group/AvatarRepository.java
@@ -1,4 +1,4 @@
-package com.alal.backend.repository.user;
+package com.alal.backend.repository.group;
 
 import com.alal.backend.domain.entity.project.Avatar;
 import org.springframework.data.jpa.repository.JpaRepository;
