+++ src/main/java/com/alal/backend/repository/group/ScriptRepository.java
@@ -1,4 +1,4 @@
-package com.alal.backend.repository;
+package com.alal.backend.repository.group;
 
 import com.alal.backend.domain.entity.project.Script;
 import org.springframework.data.jpa.repository.JpaRepository;
